+++ CHANGELOG.md
@@ -15,6 +15,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.1.0] - 2024-11-16
+
+### Added
+
+- Add comments status widget.
+
+### Changed
+
+- Optimize the way comments are stored.
+
+### Fixed
+
+- Fixed an issue where comments were lost when the file name or location changed.
+
 ## [1.0.1] - 2024-11-15
 
 ### Fixed
@@ -28,7 +42,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add comments to files and folders.
 
 [//]: # (@formatter:off)
-[unreleased]: https://github.com/conifercone/mumu-intellij-plugin/compare/v1.0.1...develop
+[unreleased]: https://github.com/conifercone/mumu-intellij-plugin/compare/v1.1.0...develop
+[1.1.0]: https://github.com/conifercone/mumu/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/conifercone/mumu/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/conifercone/mumu-intellij-plugin/releases/tag/v1.0.0
 [//]: # (@formatter:on)